EDB.Serializable: added functionality to exclude properties
When you define a method called serializable_exclude_properties in
an Object which implements the Serializable interface, you can return
a string[] of properties that are neither used for generating the
class-identifier nor for serialization itself.

This means that their value is simply ignored and does not influence
the serialization process at all.