Http.Server: fixed an issue which caused a test to fail randomly
The test http-server-session failed sometimes, but not always. Turns
out that when the response to a request - containing the neutron_session_id -
is submitted by the Entity-object, the Server-object had not yet stored
the session in its HashTable.

Created methods store_session and delete_session in the Server-Object
and which are called from the entity inside set_session which ensures
that when the client gets its session-id the session is already stored.

This renders the ServerAction-class pretty much useless.